backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.3
bitarray==3.3.0
blinker==1.9.0
boto3==1.37.27
botocore==1.37.27
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.0
click==8.1.8
cryptography==44.0.2
cytoolz==1.0.1
deepface==0.0.93
Django==5.1.7
django-cors-headers==4.7.0
django-otp==1.5.4
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
django-shortcuts==1.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dlib==19.22.99
dynaconf==3.2.10
eth-account==0.13.6
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-tester==0.12.1b1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
frozenlist==1.5.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
hexbytes==1.3.0
hvac==2.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
keras==3.9.2
knox==0.1.14
libclang==18.1.1
loguru==0.7.3
lru-dict==1.3.0
lz4==4.4.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mtcnn==1.0.0
multidict==6.2.0
namex==0.0.8
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.1.0
propcache==0.3.1
protobuf==5.29.4
psycopg2==2.9.10
py-ecc==7.0.1
py-evm==0.10.1b2
py-geth==5.4.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.9.0
pyotp==2.9.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
rlp==4.1.0
s3transfer==0.11.4
safe-pysha3==1.0.4
semantic-version==2.10.0
setuptools==78.1.0
shortuuid==1.0.13
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tf_keras==2.19.0
toolz==1.0.0
tqdm==4.67.1
trie==3.1.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
validators==0.34.0
web3==7.10.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
